backend/requirements.txt
asgiref==3.8.1
autopep8==2.2.0
diff-match-patch==20230430
Django==5.0.6
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-humanize==0.1.2
django-import-export==4.0.7
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-pgclone==3.4.1
django-taggit==5.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
humanize==4.9.0
pillow==10.3.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.5.0
tablib==3.6.1
toml==0.10.2
tzdata==2024.1
